Move getAdjustedRecordableOffencesForCourtCase into phase3 folder
