+++ packages/core/phase3/lib/getAdjustedRecordableOffencesForCourtCase.ts
@@ -1,7 +1,7 @@
-import type { Offence } from "../types/AnnotatedHearingOutcome"
+import type { Offence } from "../../types/AnnotatedHearingOutcome"
 
-import isResultCompatibleWithDisposal from "./isResultCompatibleWithDisposal"
-import getRecordableOffencesForCourtCase from "./offences/getRecordableOffencesForCourtCase"
+import isResultCompatibleWithDisposal from "../../lib/isResultCompatibleWithDisposal"
+import getRecordableOffencesForCourtCase from "../../lib/offences/getRecordableOffencesForCourtCase"
 
 const has2059Or2060Result = (offence: Offence) =>
   offence.Result.some((result) => [2059, 2060].includes(result.PNCDisposalType ?? 0))
