+++ packages/core/phase3/lib/hearingDetails/generateHearingsAndDisposals.ts
@@ -1,7 +1,7 @@
 import type { PncUpdateDataset } from "../../../types/PncUpdateDataset"
 import type { PncUpdateCourtHearingAndDisposal } from "../../types/HearingDetails"
 
-import getAdjustedRecordableOffencesForCourtCase from "../../../lib/getAdjustedRecordableOffencesForCourtCase"
+import getAdjustedRecordableOffencesForCourtCase from "../getAdjustedRecordableOffencesForCourtCase"
 import { createCourtHearingFromOffence } from "./createCourtHearingFromOffence"
 import createDisposalsFromOffence from "./createDisposalsFromOffence"
