+++ packages/core/phase3/lib/pncUpdateRequestGenerators/normalDisposalGenerator/normalDisposalGenerator.ts
@@ -5,10 +5,10 @@ import { isError } from "@moj-bichard7/common/types/Result"
 import type PncUpdateRequestGenerator from "../../../types/PncUpdateRequestGenerator"
 
 import formatDateSpecifiedInResult from "../../../../lib/createPncDisposalsFromResult/formatDateSpecifiedInResult"
-import getAdjustedRecordableOffencesForCourtCase from "../../../../lib/getAdjustedRecordableOffencesForCourtCase"
 import isResultCompatibleWithDisposal from "../../../../lib/isResultCompatibleWithDisposal"
 import { PncOperation } from "../../../../types/PncOperation"
 import generateBasePncUpdateRequest from "../../generateBasePncUpdateRequest"
+import getAdjustedRecordableOffencesForCourtCase from "../../getAdjustedRecordableOffencesForCourtCase"
 import getPncCourtCode from "../../getPncCourtCode"
 import { generateArrestHearingsAdjudicationsAndDisposals } from "../../hearingDetails/generateArrestHearingsAdjudicationsAndDisposals"
 import { generateHearingsAdjudicationsAndDisposals } from "../../hearingDetails/generateHearingsAdjudicationsAndDisposals"
